Updated tests and test folder & file structure
